cambios en pesaje, parcialidadcafe cafe, por requerimientos de examen.se edito el tipo de datos en pesaje en base de datos

+++ src/main/java/com/beneficio/beneficio/model/Cuenta.java
@@ -19,7 +19,7 @@ public class Cuenta {
     private Integer agricultor;
 
     @Column(nullable = true)
-    private Integer id_pesaje;
+    private Long id_pesaje;
 
     @Column(nullable = true)
     private Integer id_estado;
